requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
deprecation==2.1.0
discord.py==2.3.2
frozenlist==1.4.1
gotrue==2.4.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Logbook==1.7.0.post0
loguru==0.7.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
packaging==24.1
pip==24.0
postgrest==0.16.8
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
realtime==1.0.5
requests==2.32.3
rich==13.7.1
scheduler==0.8.7
six==1.16.0
sniffio==1.3.1
storage3==0.7.6
StrEnum==0.4.15
supabase==2.5.1
supafunc==0.4.6
typeguard==4.3.0
typing_extensions==4.12.2
urllib3==2.2.1
websockets==12.0
win32-setctime==1.1.0
yarl==1.9.4